// apps/cdk-apps/src/shared/domain/logging/update-create/models/ocpi-connector.ts
import { UpdatedAtDateValue } from '../../../../models/general/updated-at-date.value';
import { OcpiConnectorPrimaryId } from './ocpi-connector-primary-id';
import { OcpiEvsePrimaryIdValue } from './ocpi-evse-primary-id-value';
import { NumberValue } from './number.value';
import { isNull } from '../../../../utils/is-null.function';
import { extractPropertyIntoPayload } from '../../../../utils/extract-property-into-payload.function';
import { validatePropertyOrThrowError } from '../../../../utils/validate-property-or-throw-error';
import {
  extractAvailablePropertyIntoPayload
} from '../../../../utils/extract-available-property-into-payload.function';

export class OcpiConnector implements Readonly<OcpiConnector> {
  id: string;
  format: string;
  standard: string;
  power_type: string;
  power_kw: number;
  voltage: number;
  amperage: number;
  tariff_id: string;
  terms_and_conditions: string;
  last_updated: Date;
  primary_id: string;
  evsePrimaryId: string; // in camelCase notication because of the autogenerated column name based on OneToManyRelation
  locationId: string;
  uid: string;

  private constructor(args: Partial<OcpiConnector>) {
    Object.assign(this, args);
  }

  static createFromCreatePayload(args: Partial<OcpiConnector>): OcpiConnector {
    if (isNull(args)) {
      throw new Error(`Invalid input values for OcpiConnector.fromUpdatePayload ${JSON.stringify(args)}`);
    }

    validatePropertyOrThrowError({ properties: ['locationId', 'uid', 'id'], data: args });

    const evsePrimaryId = OcpiEvsePrimaryIdValue.create({ locationId: args.locationId!, uid: args.uid! }).value;
    const payload = extractPropertyIntoPayload({
      properties: ['id', 'standard', 'format', 'power_type', 'power_kw', 'voltage', 'amperage', 'tariff_id', 'terms_and_conditions',],
      data: args,
    });
    payload.primary_id = OcpiConnectorPrimaryId.createFromEvsePrimaryId({
      evsePrimaryId,
      connectorId: args.id!,
    }).value;
    payload.last_updated = UpdatedAtDateValue.create(args.last_updated).value;
    payload.evsePrimaryId = evsePrimaryId;
    payload.power_kw = NumberValue.create(args.power_kw).value;
    payload.voltage = NumberValue.create(args.voltage).value;
    payload.amperage = NumberValue.create(args.amperage).value;
    payload.uid = args.uid;
    payload.locationId = args.locationId;

    return new OcpiConnector(payload);
  }

  static createFromUpdatePayload(args: Partial<OcpiConnector>): OcpiConnector {
    const createPayload = this.createFromCreatePayload(args);

    const _payload = {
      ...createPayload,
      evse: {
        primary_id: createPayload.evsePrimaryId
      }
    };

    const {
      evsePrimaryId,
      ...payload
    } = extractAvailablePropertyIntoPayload(_payload);

    return new OcpiConnector(extractAvailablePropertyIntoPayload(payload));
  }
}